Move get_monitor_domains_handler to control protocol handlers
